Moved Cluster/Slave classes to separate module
