+++ src/server/server.py
@@ -38,7 +38,6 @@ import shutil
 import base64
 import subprocess
 import importlib
-import socketio
 from monotonic import monotonic
 from functools import wraps
 from collections import OrderedDict
@@ -59,6 +58,7 @@ import Language
 import RHUtils
 from RHUtils import catchLogExceptionsWrapper
 from Language import __
+from ClusterNodeSet import SlaveNode, ClusterNodeSet
 
 # Events manager
 from eventmanager import Evt, EventManager
@@ -147,199 +147,8 @@ def monotonic_to_epoch_millis(secs):
     return 1000.0*secs + MTONIC_TO_EPOCH_MILLIS_OFFSET
 
 #
-# Slaves
-#
-
-class Slave:
-
-    TIMER_MODE = 'timer'
-    MIRROR_MODE = 'mirror'
-
-    def __init__(self, idVal, info):
-        self.id = idVal
-        self.info = info
-        addr = info['address']
-        if not '://' in addr:
-            addr = 'http://'+addr
-        self.address = addr
-        self.lastContact = -1
-        self.lastCheckQueryTime = 0
-        self.freqsSentFlag = False
-        self.raceStartEpoch = 0
-        self.clockWarningFlag = False
-        self.sio = socketio.Client()
-        self.sio.on('connect', self.on_connect)
-        self.sio.on('disconnect', self.on_disconnect)
-        self.sio.on('pass_record', self.on_pass_record)
-        self.sio.on('check_slave_response', self.on_check_slave_response)
-
-    def reconnect(self):
-        if self.lastContact == -1:
-            startConnectTime = monotonic()
-            logger.info("Connecting to slave {0} at {1}...".format(self.id+1, self.address))
-            while monotonic() < startConnectTime + self.info['timeout']:
-                try:
-                    self.sio.connect(self.address)
-                    logger.info("Connected to slave {0} at {1}".format(self.id+1, self.address))
-                    gevent.spawn_later(3, start_background_threads)  # start heartbeat so 'doClusterMgmtActions()' is called
-                    return True
-                except socketio.exceptions.ConnectionError:
-                    gevent.sleep(0.1)
-            logger.warn("Connection to slave {0} at {1} failed".format(self.id+1, self.address))
-            return False
-        else:
-            return True
-
-    @catchLogExceptionsWrapper
-    def emit(self, event, data = None):
-        if self.reconnect():
-            self.sio.emit(event, data)
-            self.lastContact = monotonic()
-
-    def on_connect(self):
-        self.lastContact = monotonic()
-
-    def on_disconnect(self):
-        self.lastContact = -1
-        logger.info("Disconnected from slave {0} at {1}".format(self.id+1, self.address))
-
-    @catchLogExceptionsWrapper
-    def on_pass_record(self, data):
-        self.lastContact = monotonic()
-        node_index = data['node']
-        pilot_id = Database.HeatNode.query.filter_by( \
-            heat_id=RACE.current_heat, node_index=node_index).one_or_none().pilot_id
-
-        if pilot_id != Database.PILOT_ID_NONE:
-
-            # convert split timestamp (epoch ms sine 1970-01-01) to equivalent local 'monotonic' time value
-            split_ts = data['timestamp'] - RACE.start_time_epoch_ms
-
-            lap_count = max(0, len(RACE.get_active_laps()[node_index]) - 1)
-
-            # get timestamp for last lap pass (including lap 0)
-            act_laps_list = RACE.get_active_laps()[node_index]
-            if len(act_laps_list) > 0:
-                last_lap_ts = act_laps_list[-1]['lap_time_stamp']
-            else:
-                last_lap_ts = 0
-
-            split_id = self.id
-            last_split_id = DB.session.query(DB.func.max(Database.LapSplit.split_id)).filter_by(node_index=node_index, lap_id=lap_count).scalar()
-            if last_split_id is None: # first split for this lap
-                if split_id > 0:
-                    logger.info('Ignoring missing splits before {0} for node {1}'.format(split_id+1, node_index+1))
-                last_split_ts = last_lap_ts
-            else:
-                if split_id > last_split_id:
-                    if split_id > last_split_id + 1:
-                        logger.info('Ignoring missing splits between {0} and {1} for node {2}'.format(last_split_id+1, split_id+1, node_index+1))
-                    last_split_ts = Database.LapSplit.query.filter_by(node_index=node_index, lap_id=lap_count, split_id=last_split_id).one().split_time_stamp
-                else:
-                    logger.info('Ignoring out-of-order split {0} for node {1}'.format(split_id+1, node_index+1))
-                    last_split_ts = None
-
-            if last_split_ts is not None:
-
-                # get race-start time value from slave (or previously received and saved value)
-                slv_rs_epoch = data.get('race_start_epoch', 0) or self.raceStartEpoch
-                if slv_rs_epoch > 0:
-                    self.raceStartEpoch = slv_rs_epoch  # save race-start time for future laps
-                    # compare the local race-start epoch time to the slave's
-                    # (account for master-server's prestage race-delay time)
-                    rs_epoch_diff = RACE.start_time_epoch_ms - slv_rs_epoch - \
-                                    int((RACE.start_time_delay_secs-RACE_START_DELAY_EXTRA_SECS) * 1000)
-                    # if slave timer's clock is too far off then correct as best we can
-                    #  using estimated difference in race-start epoch times
-                    if abs(rs_epoch_diff) > 1000:
-                        split_ts += rs_epoch_diff
-                        if not self.clockWarningFlag:
-                            logger.warn("Slave {0} clock not synchronized with master, offset={1:.1f}ms, split={2}".\
-                                        format(self.id+1, (-rs_epoch_diff), split_id+1))
-                            self.clockWarningFlag = True
-                        logger.debug("Correcting slave-time offset ({0:.1f}ms) on node {1}, new split_ts={2:.1f}, split={3}".\
-                                     format((-rs_epoch_diff), node_index+1, split_ts, split_id+1))
-                    else:
-                        logger.debug("Slave {0} clock synchronized OK with master, offset={1:.1f}ms, split={2}".\
-                                     format(self.id+1, (-rs_epoch_diff), split_id+1))
-
-                split_time = split_ts - last_split_ts
-                split_speed = float(self.info['distance'])*1000.0/float(split_time) if 'distance' in self.info else None
-                split_time_str = RHUtils.time_format(split_time)
-                logger.debug('Split pass record: Node {0}, lap {1}, split {2}, time={3}, speed={4}' \
-                    .format(node_index+1, lap_count+1, split_id+1, split_time_str, \
-                    ('{0:.2f}'.format(split_speed) if split_speed is not None else 'None')))
-
-                DB.session.add(Database.LapSplit(node_index=node_index, pilot_id=pilot_id, lap_id=lap_count, \
-                        split_id=split_id, split_time_stamp=split_ts, split_time=split_time, \
-                        split_time_formatted=split_time_str, split_speed=split_speed))
-                DB.session.commit()
-                emit_current_laps() # update all laps on the race page
-                emit_phonetic_split(pilot_id, split_id, split_time);
-        else:
-            logger.info('Split pass record dismissed: Node: {0}, Frequency not defined' \
-                .format(node_index+1))
-
-    @catchLogExceptionsWrapper
-    def on_check_slave_response(self, data):
-        self.lastContact = monotonic()
-
-class Cluster:
-    def __init__(self):
-        self.slaves = []
-
-    def addSlave(self, slave):
-        slave.emit('join_cluster')
-        self.slaves.append(slave)
-
-    def hasSlaves(self):
-        return (len(self.slaves))
-
-    def emit(self, event, data = None):
-        for slave in self.slaves:
-            gevent.spawn(slave.emit, event, data)
-
-    def emitToMirrors(self, event, data = None):
-        for slave in self.slaves:
-            if slave.info['mode'] == Slave.MIRROR_MODE:
-                gevent.spawn(slave.emit, event, data)
-
-    def emitStatus(self):
-        now = monotonic()
-        SOCKET_IO.emit('cluster_status', {'slaves': [ \
-            {'address': slave.address, \
-            'last_contact': int(now-slave.lastContact) if slave.lastContact >= 0 else 'connection lost' \
-            }] for slave in self.slaves})
-        gevent.spawn(self.doClusterMgmtActions)
-
-    def doClusterMgmtActions(self):
-        now = monotonic()
-        for slave in self.slaves:
-            if not slave.freqsSentFlag:
-                try:
-                    logger.info("Sending node frequencies to slave {0} at {1}".format(slave.id+1, slave.address))
-                    slave.freqsSentFlag = True
-                    profile = getCurrentProfile()
-                    profile_freqs = json.loads(profile.frequencies)
-                    for idx in range(RACE.num_nodes):
-                        freq = profile_freqs["f"][idx]
-                        data = { 'node':idx, 'frequency':freq }
-                        slave.emit('set_frequency', data)
-                        gevent.sleep(0.001)
-                except Exception as ex:
-                    logger.error("Error sending node frequencies to slave {0} at {1}: {2}".format(slave.id+1, slave.address, ex))
-            else:
-                try:
-                    if slave.lastContact > 0 and now > slave.lastContact + 10 and now > slave.lastCheckQueryTime + 10:
-                        self.lastCheckQueryTime = now
-                        slave.sio.emit('check_slave_query')  # don't update 'lastContact' value until response received
-                except Exception as ex:
-                    logger.error("Error sending check-query to slave {0} at {1}: {2}".format(slave.id+1, slave.address, ex))
-
-#
 # Server Info
 #
-
 def buildServerInfo():
     serverInfo = {}
 
@@ -846,7 +655,7 @@ def on_load_data(data):
         elif load_type == 'vrx_list':
             emit_vrx_list(nobroadcast=True)
         elif load_type == 'cluster_status':
-            CLUSTER.emitStatus()
+            emit_cluster_status()
         elif load_type == 'hardware_log_init':
             emit_current_log_file_to_socket()
 
@@ -2877,6 +2686,14 @@ def emit_enter_and_exit_at_levels(**params):
     else:
         SOCKET_IO.emit('enter_and_exit_at_levels', emit_payload)
 
+def emit_cluster_status(**params):
+    '''Emits cluster status information.'''
+    if ('nobroadcast' in params):
+        emit('cluster_status', CLUSTER.getClusterStatusInfo())
+    else:
+        SOCKET_IO.emit('cluster_status', CLUSTER.getClusterStatusInfo())
+    gevent.spawn(CLUSTER.doClusterMgmtActions(getCurrentProfile()))
+
 def emit_start_thresh_lower_amount(**params):
     '''Emits current start_thresh_lower_amount.'''
     emit_payload = {
@@ -4010,6 +3827,10 @@ def emit_phonetic_split(pilot_id, split_id, split_time, **params):
     else:
         SOCKET_IO.emit('phonetic_split_call', emit_payload)
 
+def emit_split_pass_info(pilot_id, split_id, split_time):
+    emit_current_laps()  # update all laps on the race page
+    emit_phonetic_split(pilot_id, split_id, split_time)
+
 def emit_enter_at_level(node, **params):
     '''Emits enter-at level for given node.'''
     emit_payload = {
@@ -4191,7 +4012,7 @@ def heartbeat_thread_function():
 
             # emit cluster status less often:
             if (heartbeat_thread_function.iter_tracker % (4*HEARTBEAT_DATA_RATE_FACTOR)) == (2*HEARTBEAT_DATA_RATE_FACTOR):
-                CLUSTER.emitStatus()
+                emit_cluster_status()
 
             # collect vrx lock status
             if (heartbeat_thread_function.iter_tracker % (10*HEARTBEAT_DATA_RATE_FACTOR)) == 0:
@@ -5105,20 +4926,20 @@ if not INTERFACE or not INTERFACE.nodes or len(INTERFACE.nodes) <= 0:
         log.wait_for_queue_empty()
         sys.exit()
 
-CLUSTER = Cluster()
+CLUSTER = ClusterNodeSet()
 hasMirrors = False
 try:
     for index, slave_info in enumerate(Config.GENERAL['SLAVES']):
         if isinstance(slave_info, string_types):
-            slave_info = {'address': slave_info, 'mode': Slave.TIMER_MODE}
+            slave_info = {'address': slave_info, 'mode': SlaveNode.TIMER_MODE}
         if 'timeout' not in slave_info:
             slave_info['timeout'] = Config.GENERAL['SLAVE_TIMEOUT']
-        if 'mode' in slave_info and slave_info['mode'] == Slave.MIRROR_MODE:
+        if 'mode' in slave_info and slave_info['mode'] == SlaveNode.MIRROR_MODE:
             hasMirrors = True
         elif hasMirrors:
             logger.info('** Mirror slaves must be last - ignoring remaining slave config **')
             break
-        slave = Slave(index, slave_info)
+        slave = SlaveNode(index, slave_info, RACE, DB, start_background_threads, emit_split_pass_info)
         CLUSTER.addSlave(slave)
 except:
     logger.exception("Error adding slave to cluster")
